+++ src/index.ts
@@ -1,6 +1,8 @@
 export * from './api'
 export type {
   LayerAdapter,
+  LayerBound,
+  LayerBoundNode,
   LayerConfigFragment,
   LayerInstance,
   LayerDefine,
@@ -9,4 +11,6 @@ export type {
   LayerConfigNodeContent,
   LayerConfigContainer,
   LayerConfigCreate,
+  CloseOn,
+  CloseOnRaw,
 } from './types'